RGB: resize zones on devices
After upgrading to OpenRGB 0.8 not all the LEDs on my devices were
getting set when set_color() is called.  Apparently resizing the zones
on each device is needed prior to setting the colors.  This change
resizes each zone on each device to the number of LEDs on that device.

See https://github.com/jath03/openrgb-python/discussions/64