plugin: don't create task in the coroutine
I was wrong about my previous commit. There is a little more changes
needed for the plugin function to be a coroutine. Since the plugin is
now running in the loop we don't need to create a task of a coroutine
inside the coroutine.

Also some tests needed to be tweaked.